Add Moisture to TTN.js
This commit adds an *optional* `Moisture` data value that calculates soil moisture percentage from user-defined `adcMax` and `adcMin` variables. The default behavior is opt-in. *If un-configured,* users will receive Makerfabs default data values of *only* the battery and ADC.

Also adds `decimalPlaces` as a variable to limit data value output to two (2) decimal places by default, this variable is also configurable to suit the user's needs.

Also changes data object keys to be more descriptive. There wasn't anything wrong with them as-is but this saves time when using the Home Assistant TTN integration.